Faved_Workouts and Workout_History Table change
Before, Workout_History and Faved_Workouts were bridge tables to the Workout table. While this does make some sense, we are using the method of user defined workouts. Therefore, In order for a workout to be added to History or Favorites, It first has to be created in the Workout Table. Therefore, these tables are actually not bridge tables and should have there own bridge tables to Types

Likely won't affect much, if any of the scripted Dummy data, but will significantly affect endpoints concerning those tables and Dummy data that is manually added.