Added cardioPref Removed Suggested
I added the cardioPref attribute and removed the Suggested workouts table.  CardioPref will actually be used but I left it as nullable just in case.  Suggested Workouts was removed as the suggested workouts have been decided to be front end algorithms, and no need to store those in the DB